refactor: Lotto를 model폴더로 변경, 폴더구조 변경
- index.js파일을 이용하여 re-export
