+++ src/models/index.js
@@ -1,0 +1,3 @@
+import { default as Lotto } from './Lotto';
+
+export { Lotto };
