+++ src/utils/generateRandomNumbers.js
@@ -1,0 +1,3 @@
+const generateRandomNumbers = (min, max) => Math.floor(Math.random() * max) + min;
+
+export { generateRandomNumbers };
